refactor(filecopier): organise filecopier
re #37
