+++ src/core/usecase/FileCopier/IFileCopier.ts
